플레이리트스 삭제중

+++ resonos/src/main/java/com/cosmus/resonos/controller/PlaylistController.java
@@ -21,6 +21,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.tags.EscapeBodyTag;
 
 import com.cosmus.resonos.domain.CustomUser;
 import com.cosmus.resonos.domain.Playlist;
@@ -28,6 +29,7 @@ import com.cosmus.resonos.domain.PlaylistDTO;
 import com.cosmus.resonos.domain.Track;
 import com.cosmus.resonos.service.PlaylistDetailService;
 import com.cosmus.resonos.service.PlaylistService;
+import com.cosmus.resonos.util.UploadImage;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -81,13 +83,28 @@ public class PlaylistController {
      * 플레이리스트 생성 요청
      * @param playlist
      * @return
+     * @throws Exception
      */
-    @PostMapping
+    @PostMapping("")
     public String createPlaylist(
-        @RequestBody Playlist playlist,
-        @AuthenticationPrincipal CustomUser loginUser
-    ) {
+        @ModelAttribute Playlist playlist,
+        @AuthenticationPrincipal CustomUser loginUser,
+        @RequestParam("thumbnail") MultipartFile file
+    ) throws Exception {
+
         playlist.setUserId(loginUser.getUser().getId());
+        // 썸네일 설정 안했으면 기본 로고로
+        if(playlist.getThumbnailUrl() == null)
+            playlist.setThumbnailUrl("/img/resonosLogo.png");
+        // 이미지 파일 저장
+        if (!file.isEmpty()) {
+            UploadImage.uploadThumbnailImage(file, playlist);
+        }
+
+        boolean result = playlistService.insert(playlist);
+        if(result) {
+            return "redirect:/playlists";
+        }
 
         return "redirect:/playlists/create?fail=true";
     }
@@ -159,10 +176,8 @@ public class PlaylistController {
         return new ResponseEntity<>("트랙 삭제 실패.", HttpStatus.BAD_REQUEST);
     }
 
-
-
     /**
-     * 플레이리스트 수정
+     * 플레이리스트 수정 요청
      * @param id
      * @param playlist
      * @param loginUser
@@ -179,34 +194,7 @@ public class PlaylistController {
 
         // 이미지 파일 저장
         if (!file.isEmpty()) {
-            try {
-            String uploadDir = System.getProperty("user.dir") + "/resonos/uploads/thumbnail";
-
-            log.info("업로드 경로 : {}", uploadDir);
-
-            File folder = new File(uploadDir);
-            if (!folder.exists()) {
-                folder.mkdirs(); // 폴더 없으면 생성
-            }
-
-            // 파일 이름 중복 방지용
-            String originalFilename = file.getOriginalFilename();
-            String extension = originalFilename.substring(originalFilename.lastIndexOf("."));
-            String savedFilename = UUID.randomUUID() + extension;
-
-            // static/thumbnail 경로 (resources는 컴파일 시 target/classes로 복사됨)
-            String savePath = new File(uploadDir, savedFilename).getAbsolutePath();
-
-            // 저장
-            file.transferTo(new File(savePath));
-
-            // DB에 상대경로 저장하는 경우
-            playlist.setThumbnailUrl("/thumbnail/" + savedFilename);
-
-            log.info("이미지 저장됨");
-            } catch(Exception e) {
-                e.printStackTrace();
-            }
+            UploadImage.uploadThumbnailImage(file, playlist);
         }
 
         Playlist currentPlaylist = playlistService.select(id);
@@ -224,16 +212,31 @@ public class PlaylistController {
         return "redirect:/playlists/" + id + "?fail=true";
     }
 
+    /**
+     * 플레이리스트 삭제 요청
+     * @param id
+     * @return
+     * @throws Exception
+     */
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deletePlaylist(@PathVariable Long id) {
-        try {
-            boolean success = playlistService.delete(id);
-            if (success) {
-                return ResponseEntity.ok("Playlist deleted");
-            }
-            return ResponseEntity.status(500).body("Failed to delete playlist");
-        } catch (Exception e) {
-            return ResponseEntity.status(500).body("Failed to delete playlist: " + e.getMessage());
-        }
+    public ResponseEntity<?> deletePlaylist(
+        @PathVariable("id") Long id,
+        @AuthenticationPrincipal CustomUser loginUser
+    ) throws Exception {
+        log.info("넘어온 아이디 : {}", id);
+
+        Playlist playlist = playlistService.select(id);
+        if (playlist == null)
+            return new ResponseEntity<>("해당 플레이리스트가 존재하지 않습니다.", HttpStatus.NOT_FOUND);
+
+        Long onwnerId = playlist.getUserId();
+
+        if(!loginUser.getUser().getId().equals(onwnerId))
+            return new ResponseEntity<>("권한이 없습니다.", HttpStatus.BAD_REQUEST);
+
+        boolean result = playlistService.delete(id);
+        if(result) return new ResponseEntity<>("삭제되었습니다.", HttpStatus.OK);
+
+        return new ResponseEntity<>("삭제 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
     }
 }
